Xpdnet (#73)
Adding the xpdnet experiments (that did not reveal anything substantial but the fact that MWCNN is well adapted for MRI Reconstruction). 
Useful to have these networks already implemented as well as the multiscale complex and the XPDNet to be able to handle any new type of denoiser that arises in the future.